pyproject.toml
[tool.poetry]
name = "dcodex_collation"
version = "1.0.6"
description = "An extension for DCodex to collate transcriptions"
authors = ["Robert Turnbull <robert.turnbull@unimelb.edu.au>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/rbturnbull/dcodex_collation"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
dcodex = "^1.0.0"
django-extensions = "^3.1.3"
django-next-prev = "^1.1.0"
plotly = "^5.4.0"

[tool.poetry.dev-dependencies]
model-bakery = "^1.3.2"
coverage = "^5.5"
black = "^21.9b0"

[build-system]
requires = ["setuptools >= 40.6.0", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
